lots of updates/rewriting
STILL TO DO:

the "logic" section, which needs to be updated
significantly after all the refactoring.

The gossip section is also somewhat out of date with regards to what gossip is
actually responsible for.

Need to add a section about key->id translation. there is some stuff in the
attributes section that would probably be better handled by this functionality now.

I think we should remove a lot of the in-depth discussion of BSI and range
encoding and just point to the blog post.

The API section is one sentence which seems dumb.

once we finish updating sections the whole thing will need a re-read and
re-organize for overall coherence.